Changelist: - I tried to use async/await to avoid callback hell. Now half of my codebase is async and I have a headache. This was a terrible plan.

+++ Assets/Scripts/GameLogic/IBot.cs
@@ -12,7 +12,7 @@ namespace KarmaLogic
         public interface IBot
         {
             public string Name { get; }
-            public float DelaySeconds { get; }
+            public int DelaySeconds { get; }
             public bool IsReady(IBoard board);
             public BoardPlayerAction SelectAction(IBoard board);
             public FrozenMultiSet<CardValue> ComboToPlay(IBoard board);
